# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Finance & Crisis Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /projects/{id}/entries:
    get:
      operationId: listProjectEntries
      tags: [projects]
      summary: List all monthly entries for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of project entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectEntry"
    post:
      operationId: createProjectEntry
      tags: [projects]
      summary: Create a monthly P&L entry for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectEntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEntry"
  /projects/{id}/entries/{entryId}:
    put:
      operationId: updateProjectEntry
      tags: [projects]
      summary: Update a project entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: entryId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectEntryInput"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEntry"
    delete:
      operationId: deleteProjectEntry
      tags: [projects]
      summary: Delete a project entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: entryId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Davlatov ────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
    ProjectEntry:
      type: object
      required: [id, projectId, month, grossRevenue, directCosts, marketingExpense, salaryExpense, rentExpense, logisticsExpense, utilitiesExpense, reinvestment, dividends, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        month:
          type: string
        grossRevenue:
          type: number
        directCosts:
          type: number
        marketingExpense:
          type: number
        salaryExpense:
          type: number
        rentExpense:
          type: number
        logisticsExpense:
          type: number
        utilitiesExpense:
          type: number
        reinvestment:
          type: number
        dividends:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectEntryInput:
      type: object
      required: [month]
      properties:
        month:
          type: string
        grossRevenue:
          type: number
        directCosts:
          type: number
        marketingExpense:
          type: number
        salaryExpense:
          type: number
        rentExpense:
          type: number
        logisticsExpense:
          type: number
        utilitiesExpense:
          type: number
        reinvestment:
          type: number
        dividends:
          type: number
        notes:
          type: string
...